backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.4.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
httpx==0.25.2
pytest==7.4.3
requests==2.31.0
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0